social-flow-backend/requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Redis
redis[hiredis]==5.0.1
aioredis==2.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pyotp==2.9.0
qrcode[pil]==7.4.2
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
authlib==1.3.0

# Validation
pydantic[email]==2.5.0
email-validator==2.1.0

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# AWS SDK
boto3==1.34.0
botocore==1.34.0
aioboto3==12.3.0

# Background Tasks
celery==5.3.4
redis==5.0.1

# Monitoring & Logging
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
structlog==23.2.0

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78

# Video Processing
ffmpeg-python==0.2.0

# ML & AI
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
tensorflow==2.15.0
torch==2.1.1
transformers==4.36.0
opencv-python==4.8.1.78

# Search
elasticsearch==8.11.0
elasticsearch-dsl==8.11.0

# Payments
stripe==7.8.0

# Email
aiosmtplib==3.0.1
jinja2==3.1.2

# WebSockets
websockets==12.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Code Quality
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Development
python-dotenv==1.0.0
